finish search

+++ src/main/java/me/skynda/classification/dto/SliderAttributesDto.java
@@ -2,8 +2,10 @@ package me.skynda.classification.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class SliderAttributesDto extends ClassificationBaseDto {
-    public int Min;
-    public int Max;
+    public Double Min;
+    public Double Max;
 }
